requirements.txt
boto3==1.28.3
botocore==1.31.3
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
Django==4.2.3
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-countries==7.5.1
django-elasticsearch==7.0.2
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-filter==23.2
django-js-asset==2.1.0
django-minio-backend==3.5.0
django-modeltranslation==0.18.11
django-mptt==0.14.0
django-nine==0.2.7
django-redis==5.3.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.6
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
fastapi==0.101.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
Jinja2==3.1.2
minio==7.1.15
oauthlib==3.2.2
PaycomUz==3.0
Pillow==10.0.0
pip-chill==1.0.3
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pyflakes==3.0.1
PyJWT==2.7.0
pyOpenSSL==23.2.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
SQLAlchemy==2.0.19
sqlparse==0.4.4
starlette==0.27.0
ujson==5.8.0
uritemplate==4.1.1
uvicorn==0.23.2
